+++ TP10_AgenceAuto/src/agence/Vehicule.java
@@ -1,3 +1,6 @@
+/** 
+ * @author Jeremy Fouquet
+ */
 package agence;
 
 import java.text.SimpleDateFormat;
@@ -9,27 +12,52 @@ import java.util.Objects;
 import exception.MonException;
 
 public abstract class Vehicule {
-	private TypeMoteur carburant;
-	private String immatriculation;
-	private double prix;
-	private int nbroues;
-	private double nbkilometres = 0;
-	private double louejusque = Double.NaN;
-	private Date dateCT;
-	private Date dateMiseEnService;
-	private List<Moteur> moteurs = new ArrayList<Moteur>(2);
+	private TypeMoteur carburant; // type de moteur (Electrique, Essence, Hybride)
+	private String immatriculation; // immatriculation unique à chaque vehicule
+	private double prix; // prix du vehicule
+	private int nbroues; // nombre de roues du vehicule
+	private double nbkilometres = 0; // nombre total de kilometre parcouru par le vehicule depuis la mise en circulation
+	private double louejusque = Double.NaN; // Dans le cas d'une location la valeur ne doit jamais être dépassé par this.nbkilometres
+	private Date dateCT; // Coorespond à la derniere date du contrôle technique realisé sinon null
+	private Date dateMiseEnService; // correspond à la date d'instance de classe
+	private List<Moteur> moteurs = new ArrayList<Moteur>(2); // Liste des moteurs du vehicule s'il en dispose (cf:carburant)
 
-	public Vehicule(int prix) {
+	/**
+	 * 
+	 * Constructeur en cas de vehicule mecanique
+	 * 
+	 * @param prix
+	 */
+	public Vehicule(double prix) {
 		this.immatriculation = this.getClass().getSimpleName()+AgenceAuto.id++;
 		dateMiseEnService = new Date();
 	}
 
-	public Vehicule(int prix, TypeMoteur typeMoteur) {
+	/**
+	 * 
+	 * Constructeur en cas de vehicule à moteur
+	 * 
+	 * @param prix
+	 * @param typeMoteur
+	 */
+	public Vehicule(double prix, TypeMoteur typeMoteur) {
 		this(prix);
 	}
 	
+	/**
+	 * 
+	 * Retourne une String correspondant au klaxonne du vehicule
+	 * 
+	 * @return String klaxonne du vehicule
+	 */
 	public abstract String klaxonner();
 	
+	/**
+	 * 
+	 * Calcule le nombre de kilometre que le vehicule est capable de faire en fonction de son moteur
+	 * 
+	 * @return double nombre de kilometre pouvant être réalisé par le vehicule infinity si le vehicule est mecanique
+	 */
 	public double nbKilometreRestant() {
 		double km = Double.POSITIVE_INFINITY;
 		if(this.moteurs.size() > 0) {
@@ -49,6 +77,14 @@ public abstract class Vehicule {
 		return km;
 	};
 	
+	/**
+	 * 
+	 * Ajout unite par unite une quantite souhaité de carburant au reservoire du vehicule s'il en a jusqu'à ce qu'il soit plein
+	 * 
+	 * @param quantite quantite de carburant souhaité ajouté au vehicule
+	 * @param carburant Electrique ou Essence
+	 * @exception MonException Si le carburant ne convient à aucun moteur
+	 */
 	public void ajouterCarburant(double quantite, Carburant carburant) throws MonException {
 		if(this.moteurs.size() > 0) {
 			double ajoute = quantite;
@@ -81,17 +117,31 @@ public abstract class Vehicule {
 	};
 	
 	
-	
+	/**
+	 * 
+	 * Affiche le nombre de kilometre parcouru au total avec le vehicule
+	 * 
+	 */
 	public void kilometreParcouru() {
 		System.out.printf("%s %.1f %s\n", "Le vehicule a parcouru", this.nbkilometres, "kilomètres");
 	}
 	
+	/**
+	 * 
+	 * Retrourne le vehicule loué parmis les vehicule à vendre
+	 */
 	public void rendreLocation() {
 		AgenceAuto.vehiculesLoues.remove(this);
 		AgenceAuto.vehicules.add(this);
 		this.louejusque = Double.NaN;
 	}
-	
+
+	/**
+	 * 
+	 * Parcoure km par km une distance souhaité (pour les vehicule à moteur : tant que le reservoir de carburant n'est pas vide)
+	 * 
+	 * @param kilometresAParcourir distance à parcourir
+	 */
 	public void seDeplacer(double kilometresAParcourir) {
 		double avance = 1;
 		double parcouru = kilometresAParcourir;
@@ -121,7 +171,13 @@ public abstract class Vehicule {
 		
 		
 	}
-	
+
+	/**
+	 * 
+	 * Affiche un message d'information sur la date du prochain contrôle technique et calcule l'annee du prochain controle
+	 * 
+	 * @return int distance en années (à aujourd'hui) jusqu'au prochain contrôle technique
+	 */
 	public int prochainCT() {
 		Date aujourdhui = new Date();
 		int annees = 0;
@@ -144,15 +200,33 @@ public abstract class Vehicule {
 		return annees;
 	}
 	
+	/**
+	 * 
+	 * Affecte la date du contrôle technique à aujourd'hui
+	 * 
+	 */
 	public void faireCT() {
 		Date aujourdhui = new Date();
 		this.dateCT = aujourdhui;
 	}
 	
+	/**
+	 * 
+	 * loue le vehicule pour un certain nombre de kilometre
+	 * 
+	 * @param double nombre de kilometre durant lequels la location est valide
+	 */
 	protected void loue(double nbKilometresLoue) {
 		this.louejusque = nbkilometres+nbKilometresLoue;
 	}
 	
+	/**
+	 * 
+	 * Selectionne l'un des moteur de la liste du vehicule en fonction de son carburant
+	 * 
+	 * @param carburant permet de distinguer le moteur souhaité
+	 * @return Moteur moteur souhaité
+	 */
 	private Moteur trouveLeMoteurPourLeBonCarburant(Carburant carburant) {
 		List<Moteur> moteur = new ArrayList<Moteur>(1);
 		for (Moteur m : this.moteurs) {
@@ -167,12 +241,25 @@ public abstract class Vehicule {
 		return moteurTrouve;
 	}
 
+	/**
+	 * 
+	 * Convertie un Date en int à partir de son année
+	 * 
+	 * @param date date depuis laquelle est recupéré l'année
+	 * @return int année en int de l'année passé en paramettre
+	 */
 	private int convertionDateEnAnnees(Date date) {
 		SimpleDateFormat df = new SimpleDateFormat("yyyy");
 		int convertion = Integer.parseInt(df.format(date));
 		return convertion;
 	}
-
+	
+	/**
+	 * 
+	 * Verifie qu'il y ait 1 moteur et du carburant dans celui ci
+	 * 
+	 * @return boolean si le moteur à du carburant true sinon false
+	 */
 	private boolean moteurPremierActif() {
 		boolean toReturn = false;
 		if (moteurs.size() > 0) {
@@ -183,6 +270,12 @@ public abstract class Vehicule {
 		return toReturn;
 	}
 	
+	/**
+	 * 
+	 * Verifie qu'il y ait 2 moteurs et du carburant dans le 2eme
+	 * 
+	 * @return boolean si le 2eme moteur à du carburant true sinon false
+	 */
 	private boolean moteurSecondActif() {
 		boolean toReturn = false;
 		if (moteurs.size() > 1) {
@@ -193,6 +286,12 @@ public abstract class Vehicule {
 		return toReturn;
 	}
 	
+	/**
+	 * 
+	 * Verifie si la location est terminé
+	 * 
+	 * @return boolean si la location est terminé true, dans tous les autres cas false
+	 */
 	private boolean doitEtreRendu() {
 		boolean toReturn = false;
 		if(enCoursDeLocation()) {
